requirements.txt
flask
gunicorn
numpy
matplotlib.pyplot
requests
ast